Move the sphinx build bots into standard_builders.
Now the sphinx build bots will be treated like the others
so devs will now be notified if they break the sphinx builds.

Before the sphinx build bots were only notifying Sean Silva and myself.
This was so we could evaluate whether or not the bots would be too
noisy. They don't seem to be particularly noisy so it makes sense to
just have everyone notified when build failures occur.