Disable SingleSource/UnitTests/blockstret on AArch64.
It never worked, and isn't supposed to.

It tests that clang properly emits the BLOCK_USE_STRET bit in the block
infos.  But that's intentionally disabled in:
  AArch64TargetCodeGenInfo::doesReturnSlotInterfereWithArgs
which indirectly drives the BLOCK_USE_STRET decision, via:
  CodeGenModule::ReturnSlotInterferesWithArgs

On AArch64, sret parameters go through a dedicated register (x8), and
thus don't affect the ABI like they would elsewhere (where all "actual"
parameters would be forced to be passed in the next register defined by
the ABI).

So, on AArch64, from the runtime's standpoint, blocks never "use STRET".